\name{admisc-internal}
\alias{checkMV}
\alias{prettyTable}
\alias{dashes}
\alias{splitstr}
\alias{tildae}
\alias{trimstr}

\title{Internal Functions}

\description{
  The above functions are internal in the admisc package which are not designed to
  be called directly by the user.
}

\keyword{internal}
