% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admisc_package.R
\docType{package}
\name{admisc-package}
\alias{admisc}
\alias{admisc-package}
\title{admisc: Adrian Dusa's Miscellaneous}
\description{
Contains functions used across packages 'DDIwR', 'QCA' and 'venn'. Interprets and translates, factorizes and negates SOP - Sum of Products expressions, for both binary and multi-value crisp sets, and extracts information (set names, set values) from those expressions. Other functions perform various other checks if possibly numeric (even if all numbers reside in a character vector) and coerce to numeric, or check if the numbers are whole. It also offers, among many others, a highly versatile recoding routine and some more flexible alternatives to the base functions 'with()' and 'within()'. SOP simplification functions in this package use related minimization from package 'QCA', which is recommended to be installed despite not being listed in the Imports field, due to circular dependency issues.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/dusadrian/admisc}
  \item Report bugs at \url{https://github.com/dusadrian/admisc/issues}
}

}
\author{
\strong{Maintainer}: Adrian Dusa \email{dusa.adrian@unibuc.ro} (\href{https://orcid.org/0000-0002-3525-9253}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Adrian Dusa \email{dusa.adrian@unibuc.ro} (\href{https://orcid.org/0000-0002-3525-9253}{ORCID}) [copyright holder]
}

}
\keyword{internal}
