\name{admisc-package}

\alias{admisc-package}

\alias{admisc}

\docType{package}

\title{
Adrian Dusa's Miscellaneous
}

\description{
Contains functions used across packages 'QCA', 'DDIwR', and 'venn'.
Interprets and translates DNF - Disjunctive Normal Form and SOP - Sum of
Products expressions, for both binary and multi-value crisp sets, and extracts
information (set names, set values) from those expressions. Other functions
perform various other checks if possibly numeric (even if all numbers reside
in a character vector) and coerce to numeric, or check if the numbers are
whole. It also offers, among others, a highly flexible recoding function.
Some of the functions in this package use related functions from package
\bold{\pkg{QCA}}, and users are encouraged to install that package despite not
being listed in the Imports field (due to circular dependency issues).
}

\details{
    \tabular{ll}{
    Package: \tab admisc\cr
    Type:    \tab Package\cr
    Version: \tab 0.11\cr
    Date:    \tab 2020-10-04\cr
    License: \tab GPL (>= 2)\cr
    }
}

\author{

\bold{Authors}:\cr
Adrian Dusa\cr
Department of Sociology\cr
University of Bucharest\cr
\email{dusa.adrian@unibuc.ro}

\bold{Maintainer}:\cr
Adrian Dusa
}


\keyword{package}

