% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betweenQuotes.R
\name{betweenQuotes}
\alias{betweenQuotes}
\title{Extract information between quotes in a string}
\description{
Functions to extract the between the (escaped) quotes, in a string.
}
\usage{
betweenQuotes(x)
}

\arguments{
  \item{x}{A string.}
}


\author{
Adrian Dusa
}

\examples{
x <- "An example of \"quoted\" text."

betweenQuotes(x)
}

\keyword{functions}
