\name{scan.clipboard}

\alias{scan.clipboard}
\alias{write.clipboard}

\title{Cross platform scan/write clipboard}

\description{
Functions to read and write to the system's clipboard, for copy/paste operations.
}

\usage{
scan.clipboard(...)
write.clipboard(x)
}

\arguments{
    \item{x}{Object to be written to the clipboard}
    \item{...}{Same arguments that are used in the base function \bold{\code{scan}}}
}


\author{
Adrian Dusa
}

\keyword{functions}
