% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permutations.R
\name{permutations}
\alias{permutations}
\title{Calculates the permutations of a vector}
\description{
Generates all possible permutations of elements from a vector.
}
\usage{
permutations(x)
}

\arguments{
    \item{x}{Any kind of vector.}
}


\author{
Adrian Dusa
}

\examples{

permutations(1:3)

}

\keyword{functions}
