\name{scan.clipboard}

\alias{scan.clipboard}

\title{Cross platform scan clipboard}

\description{
A first attempt.
}

\usage{
scan.clipboard(...)
}

\arguments{
    \item{...}{Same arguments that are used in the base function \bold{\code{scan}}}
}


\author{
Adrian Dusa
}

\keyword{functions}
