\name{admisc-internal}
\alias{checkMV}
\alias{checkSubset}
\alias{prettyTable}
\alias{dashes}
\alias{splitstr}
\alias{tildae}
\alias{trimstr}
\alias{expandBrackets}
\alias{getBigList}
\alias{splitMainComponents}
\alias{splitBrackets}
\alias{removeSingleStars}
\alias{splitPluses}
\alias{splitStars}
\alias{splitTildas}
\alias{solveBrackets}
\alias{sortExpressions}
\alias{simplifyList}
\alias{negateValues}
\alias{removeDuplicates}
\alias{getNonChars}
\alias{splitProducts}

\title{Internal Functions}

\description{
  The above functions are internal in the admisc package which are not designed to
  be called directly by the user.
}

\keyword{internal}
