\name{admisc-internal}
\alias{checkMV}
\alias{checkSubset}
\alias{classify}
\alias{dashes}
\alias{expandBrackets}
\alias{getInfo}
\alias{getLevels}
\alias{getMatrix}
\alias{getNonChars}
\alias{getInfo}
\alias{getLevels}
\alias{negateLoop}
\alias{prettyString}
\alias{prettyTable}
\alias{reload}
\alias{removeSingleStars}
\alias{splitMainComponents}
\alias{splitstr}
\alias{splitBrackets}
\alias{splitPluses}
\alias{splitProducts}
\alias{splitStars}
\alias{splitTildas}
\alias{solveBrackets}
\alias{sortExpressions}
\alias{simplifyList}
\alias{tildae}
\alias{trimstr}
\alias{uninstall}
\alias{unload}
\alias{unlockEnvironment}
\alias{validateNames}
\alias{verify}
\alias{writePrimeimp}

\alias{names_values}
\alias{sort_labelled}
\alias{unique_labelled}
\alias{to_labels}
\alias{tagged_na_value}
\alias{get_na_tag}
\alias{make_tagged_na}
\alias{tag_from_string}
\alias{tagged_string}


\title{Internal Functions}

\description{
  The above functions are internal in the admisc package which are not designed to
  be called directly by the user.
}

\keyword{internal}
