% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admisc_internal.R
\name{admisc_internal}
\alias{admisc_internal}
\alias{anyTagged}
\alias{checkMV}
\alias{checkSubset}
\alias{classify}
\alias{dashes}
\alias{doublequotes}
\alias{expandBrackets}
\alias{getInfo}
\alias{getLevels}
\alias{getMatrix}
\alias{getNonChars}
\alias{getTag}
\alias{hasTag}
\alias{makeTag}
\alias{negateLoop}
\alias{padLeft}
\alias{padRight}
\alias{padBoth}
\alias{prettyString}
\alias{prettyTable}
\alias{reload}
\alias{removeSingleStars}
\alias{splitMainComponents}
\alias{splitstr}
\alias{splitBrackets}
\alias{splitPluses}
\alias{splitProducts}
\alias{splitStars}
\alias{splitTildas}
\alias{solveBrackets}
\alias{sortExpressions}
\alias{simplifyList}
\alias{singlequotes}
\alias{spaces}
\alias{stopError}
\alias{tildae}
\alias{trimstr}
\alias{uninstall}
\alias{unload}
\alias{checkValid}
\alias{validateNames}
\alias{verify}
\alias{writePIs}
\alias{writePrimeimp}
\title{admisc internal functions}
\description{
Functions to be used internally in package \code{admisc}.
}
\keyword{internal}
