\name{admisc internal functions}
\alias{anyTagged}
\alias{checkMV}
\alias{checkSubset}
\alias{classify}
\alias{dashes}
\alias{doublequotes}
\alias{expandBrackets}
\alias{getInfo}
\alias{getLevels}
\alias{getMatrix}
\alias{getNonChars}
\alias{getInfo}
\alias{getLevels}
\alias{getTag}
\alias{hasTag}
\alias{makeTag}
\alias{negateLoop}
\alias{padLeft}
\alias{padRight}
\alias{padBoth}
\alias{prettyString}
\alias{prettyTable}
\alias{reload}
\alias{removeSingleStars}
\alias{splitMainComponents}
\alias{splitstr}
\alias{splitBrackets}
\alias{splitPluses}
\alias{splitProducts}
\alias{splitStars}
\alias{splitTildas}
\alias{solveBrackets}
\alias{sortExpressions}
\alias{simplifyList}
\alias{singlequotes}
\alias{spaces}
\alias{stopError}
\alias{tildae}
\alias{trimstr}
\alias{uninstall}
\alias{unload}
\alias{checkValid}
\alias{validateNames}
\alias{verify}
\alias{writePrimeimp}
\alias{writePIs}


\title{Internal Functions}

\description{
  The above functions are internal in the admisc package which are not designed to
  be called directly by the user.
}

\keyword{internal}
